+++ src/dbapi/driver/public.cpp
@@ -459,7 +459,7 @@ bool CDB_Connection::IsReusable() const
     return m_ConnImpl->IsReusable();
 }
 
-size_t CDB_Connection::GetReuseCount() const
+unsigned int CDB_Connection::GetReuseCount() const
 {
     CHECK_CONNECTION(m_ConnImpl);
     return m_ConnImpl->GetReuseCount();
